refactor: separate OMG.Watcher.Supervisor out of Application
